# optimize the Batch.*Deferred methods