Remove the unnecessary `*WriteOptions` parameter. That parameter is
necessary for the non-deferred methods to adhere to the `Writer`
interface.

Remove the `error` return value from the `*Deferred` ops. This was done
by changing the type of `Batch.count` from a `uint32` to a `uint64` and
checking for overflow when the count is set in `Batch.Repr()` (or really
any time `Batch.Count()` is called).

Manually inline encoding of varint32s. Simplify some of the slice math
in `Batch.prepareDeferred*`.

```
name                 old time/op  new time/op  delta
BatchSet-16          22.9ns ± 1%  15.5ns ± 2%  -32.26%  (p=0.000 n=10+10)
BatchSetDeferred-16  22.1ns ± 4%  13.7ns ± 1%  -37.85%  (p=0.000 n=10+9)
```