@petermattis 2019-10-08
This came out of investigation into the `BenchmarkBatchBuilderPut` regression in the CRDB repo.